crawler/requirements.txt
argparse
coloredlogs==10.0
requests==2.20.1
requests-html==0.10.0
python-magic>=0.4.15
urllib3==1.22
redis==3.2.1